+++ sample/prelude-modules.el
@@ -79,6 +79,7 @@
 (require 'prelude-js)
 ;; (require 'prelude-latex)
 (require 'prelude-lisp) ;; Common setup for Lisp-like languages
+;; (require 'prelude-literate-programming) ;; Setup for Literate Programming
 (require 'prelude-lsp) ;; Base setup for the Language Server Protocol
 ;; (require 'prelude-lua)
 ;; (require 'prelude-ocaml)